refactor: rewrite files to TypeScript
